Moved the pure python part into type.dto.pure and the alchemy dependant one into type.dto. This way the change is completely transparent to the library users
